requirements.txt
anyio==3.5.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pydantic==1.9.0
pymongo==4.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
Werkzeug==2.1.1
WTForms==3.0.1
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
